### 2024-03-19 3c2aef5
explicitely implement Mention eq
