### 2024-03-25 #6505 @Aethor
In the meanwhile, we can suggest some reviewers from the JOSS list. Since the software is about NLP and network science, maybe one reviewer from each discipline would be appropriate. On the NLP side, Börje F. Karlsson may be interested, while Louis Boucherie may be interested on the network side.
