### 2024-03-25 f2fb691
The "chapters" feature has been renamed into "dynamic_blocks"
