+++ renard/pipeline/graph_extraction.py
@@ -19,61 +19,63 @@ def sent_index_for_token_index(token_index: int, sentences: List[List[str]]) -> 
     return next((i for i, l in enumerate(sents_len) if l > token_index))
 
 
-def sent_indices_for_chapter(
-    chapters: List[List[str]], chapter_idx: int, sentences: List[List[str]]
+def sent_indices_for_blocks(
+    dynamic_blocks: List[List[str]], block_idx: int, sentences: List[List[str]]
 ) -> Tuple[int, int]:
     """Return the indices of the first and the last sentence of a
-    chapter
+    block
 
-    :param chapters: all chapters
-    :param chapter_idx: index of the chapter for which sentence
+    :param dynamic_blocks: all blocks
+    :param block_idx: index of the block for which sentence
         indices are returned
     :param sentences: all sentences
     :return: ``(first sentence index, last sentence index)``
     """
-    chapter_start_idx = sum([len(c) for i, c in enumerate(chapters) if i < chapter_idx])
-    chapter_end_idx = chapter_start_idx + len(chapters[chapter_idx])
+    block_start_idx = sum(
+        [len(c) for i, c in enumerate(dynamic_blocks) if i < block_idx]
+    )
+    block_end_idx = block_start_idx + len(dynamic_blocks[block_idx])
     sents_start_idx = None
     sents_end_idx = None
     count = 0
     for sent_i, sent in enumerate(sentences):
         start_idx, end_idx = (count, count + len(sent))
         count = end_idx
-        if sents_start_idx is None and start_idx >= chapter_start_idx:
+        if sents_start_idx is None and start_idx >= block_start_idx:
             sents_start_idx = sent_i
-        if sents_end_idx is None and end_idx >= chapter_end_idx:
+        if sents_end_idx is None and end_idx >= block_end_idx:
             sents_end_idx = sent_i
             break
     assert not sents_start_idx is None and not sents_end_idx is None
     return (sents_start_idx, sents_end_idx)
 
 
-def mentions_for_chapters(
-    chapters: List[List[str]],
+def mentions_for_blocks(
+    dynamic_blocks: List[List[str]],
     mentions: List[Tuple[Character, NEREntity]],
 ) -> List[List[Tuple[Character, NEREntity]]]:
-    """Return each chapter mentions
+    """Return each block mentions
 
-    :param chapters:
+    :param blocks:
     :param mentions:
 
-    :return: a list of mentions per chapters.  This list has len
-             ``len(chapters)``.
+    :return: a list of mentions per blocks.  This list has len
+             ``len(blocks)``.
     """
-    chapters_mentions = [[] for _ in range(len(chapters))]
+    blocks_mentions = [[] for _ in range(len(dynamic_blocks))]
 
     start_indices = list(
-        itertools.accumulate([0] + [len(chapter) for chapter in chapters[:-1]])
+        itertools.accumulate([0] + [len(block) for block in dynamic_blocks[:-1]])
     )
-    end_indices = start_indices[1:] + [start_indices[-1] + len(chapters[-1])]
+    end_indices = start_indices[1:] + [start_indices[-1] + len(dynamic_blocks[-1])]
 
     for mention in mentions:
-        for chapter_i, (start_i, end_i) in enumerate(zip(start_indices, end_indices)):
+        for block_i, (start_i, end_i) in enumerate(zip(start_indices, end_indices)):
             if mention[1].start_idx >= start_i and mention[1].end_idx < end_i:
-                chapters_mentions[chapter_i].append(mention)
+                blocks_mentions[block_i].append(mention)
                 break
 
-    return chapters_mentions
+    return blocks_mentions
 
 
[…]
