### 2024-03-25 1bbf19d
The "chapters" feature has been renamed into "dynamic_blocks"
