### 2024-03-25 b4f2752
Make it so steps can modify some pipeline parameters if needed
