### 2024-03-25 82bcd99
Make it so steps can modify some pipeline parameters if needed
