### 2024-03-25 8bd0df4
crude support for extracting graphs with other types of entities
