### 2024-03-25 224ac6b
crude support for extracting graphs with other types of entities
