### 2024-03-29 e2e2bcd
change CI to only run on the main branch
