+++ .github/workflows/ci.yml
@@ -3,6 +3,7 @@ on: [push]
 jobs:
   ci:
     runs-on: ubuntu-latest
+    if: github.ref == 'refs/head/main'
     steps:
       - uses: actions/checkout@v2
       - uses: actions/setup-python@v2
