### 2024-03-29 3a273ff
fix wrong CI condition check
