### 2024-04-01 #6505 @logological
@tellarin, @flekova, and @LCB0B, would you be willing and able to review this submission to the Journal of Open Source Software?  (In case you aren't familiar with the journal or the reviewing process, information can be found on [the JOSS website](https://joss.theoj.org/) and [Reviewer Guide](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).)
