### 2024-04-02 #6574 @editorialbot
[REVIEW]: Renard: A Modular Pipeline for Extracting Character Networks from Narrative Texts
**Submitting author:** @Aethor (<a href="http://orcid.org/0000-0003-4629-0923">Arthur Amalvy</a>)
**Repository:** https://github.com/CompNet/Renard
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.4.2
