### 2024-04-09 b149ac0
fix a rare crash in character unification
