### 2024-05-09 #6574 @logological
@tellarin Thanks for confirming your participation.  If you could please generate your checklist as described at the top of this issue, and start evaluating the criteria in it, that would be great!  In the meantime I've sent @LCB0B another ping, this time by e-mail.
