### 2024-05-13 #6574 @LCB0B
### Functionality

 **Installation:** maybe add python version in readme (>3.8, <3.11)

 **Functionality:** Have the functional claims of the software been confirmed?
- There might be some author disambiguation issues,  i.e. figure1 shows a character name "mr", moreover there are a 'mr bingley"and 'mrs bingley" but also just 'bingley"

### Documentation
 **A statement of need:** Target audience is not clearly stated
 **Example usage:** The example is not in the documentation but in /renard_tutorial.py (the example is also mentioned in the readme, could be nice to add it to the documentation)

 **Automated tests:** Are there automated tests or manual steps described so that the functionality of the software can be verified? (still checking)

### Software paper

  **Quality of writing:**
minor typos:
- Line 47: The word "examplifies" should be corrected to "exemplifies."
- Line 50 : (at the end) two characterS
- Line 83 : should be "Rochat, Y." for consistencies between references
