### 2024-05-13 dad509d
Add basic example to README
