### 2024-05-13 b52d9a1
Add supported Python versions in README
