### 2024-05-13 2412971
fix typos and a reference issue in the JOSS article
