+++ paper.bib
@@ -61,7 +61,7 @@
 
 @PhdThesis{rochat-2014-phd_thesis,
   school       = {Université de Lausanne},
-  author       = {Rochat, Yannick},
+  author       = {Rochat, Y.},
   year	       = {2014},
   month	       = {12},
   title	       = {Character Networks and Centrality},
