+++ _sources/reference.rst.txt
@@ -184,10 +184,3 @@ Plot utils
 
 .. automodule:: renard.plot_utils
    :members:
-
-
-NER utils
----------
-
-.. automodule:: renard.ner_utils
-   :members:
