### 2024-05-13 76beb7d
JOSS article: digitial humanity -> digital humanities
