### 2024-05-13 #6574 @Aethor
Hi,

thank you for taking the time to review our article and software!

>Installation: maybe add python version in readme (>3.8, <3.11)

This is a good point, we added it in the README (see [b52d9a1c942d15f4431d752002e7bc6334c6f194](https://github.com/CompNet/Renard/commit/b52d9a1c942d15f4431d752002e7bc6334c6f194)).

>Functionality: Have the functional claims of the software been confirmed?

>	There might be some author disambiguation issues, i.e. figure1 shows a character name "mr", moreover there are a 'mr bingley"and 'mrs bingley" but also just 'bingley"

Renard is (unfortunately!) bound to make some errors. We base our alias resolution algorithm on Vala et al., 2015 (https://aclanthology.org/D15-1088/), but the problem is not trivial and not solved in general. We hope to improve alias resolution as development goes further.

>A statement of need: Target audience is not clearly stated

Good point. We added a statement of need in the article itself and in the documentation (https://compnet.github.io/Renard/introduction.html#target-audience-and-intended-usage).

>Example usage: The example is not in the documentation but in /renard_tutorial.py (the example is also mentioned in the readme, could be nice to add it to the documentation)

We have an example in the documentation (see https://compnet.github.io/Renard/pipeline.html) but not in the README. We guess this is what you mean here? Since it is useful, we added the basic example from the documentation in the README (see https://github.com/CompNet/Renard/commit/dad509d424fbecd67fc2766cff262dc10cc6faaf).

>Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified? (still checking)

Yes! You can see this under the "Running Tests" section of the README, or in the Contributing section of the doc where we highlight how to use tests (https://compnet.github.io/Renard/contributing.html). All tests are located under ./tests. We use pytest and hypothesis to help testing, and we have a CI in place on the main branch to check for issues. The CI also automatically builds the documentation.

>Quality of writing:
>minor typos:

>    Line 47: The word "examplifies" should be corrected to "exemplifies."
>    Line 50 : (at the end) two characterS
>    Line 83 : should be "Rochat, Y." for consistencies between references

Good catch, we fixed these and updated the article.
