### 2024-05-15 #6574 @LCB0B
Thank you for the quick reply, I only included the points where I still have questions/remarks. I checked the other points.

> > Functionality: Have the functional claims of the software been confirmed?
>
> > There might be some author disambiguation issues, i.e. figure1 shows a character name "mr", moreover there are a 'mr bingley"and 'mrs bingley" but also just 'bingley"
>
> Renard is (unfortunately!) bound to make some errors. We base our alias resolution algorithm on Vala et al., 2015 (https://aclanthology.org/D15-1088/), but the problem is not trivial and not solved in general. We hope to improve alias resolution as development goes further.
>
indeed author disambiguation is a hard problem, but regarding "titles" (like Mr.) , isn't it what renard/resources/titles/titles.py is suppose to do ?

> > Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified? (still checking)
>
All the tests run perfectly except that the  .hypothesis/ folder, that is need for running some tests, is included in the .gitignore
