### 2024-05-15 #6574 @Aethor
>indeed author disambiguation is a hard problem, but regarding "titles" (like Mr.) , isn't it what renard/resources/titles/titles.py is suppose to do ?

Currently, titles help to determine the gender of characters, and to correctly parse their name. If a lone title (such as the token "Mister" or "Mr.") is detected as a full entity by the NER step, it will be considered as a character in its own right. Sometimes this is actually needed, because some characters may be referred by a title exclusively or almost exclusively (this happens for example with "the Director" in Huxley's "Brave New World", or "the Lady" in Glen Cook's fantasy novel "The Black Company"). However, it may be a good idea to add a parameter to the character unification module to force Renard to ignore such instances of lone titles, which makes sense for a good number of novels. What do you think of such a solution?

>All the tests run perfectly except that the .hypothesis/ folder, that is needed for running some tests, is included in the .gitignore

From what I understand, Hypothesis tests should run without the `.hypothesis` folder. Given the hypothesis documentation, I expect that this folder contains already tested examples as a cache. Since the doc explains that the default usecase is to not track the folder under version control (https://hypothesis.readthedocs.io/en/hypothesis-python-4.57.1/database.html#sharing-your-example-database), we did not do it, but there might be good reasons to do so?
