### 2024-05-16 95385d1
Add the possibility to use a custom co-occurrences window
