### 2024-05-16 647ab67
Add the possibility to use a custom co-occurrences window
