+++ renard/pipeline/graph_extraction.py
@@ -1,11 +1,10 @@
-import itertools
 from typing import Dict, Any, List, Set, Optional, Tuple, Literal, Union
+import itertools as it
 import operator
-from itertools import accumulate
 
 import networkx as nx
 import numpy as np
-from more_itertools import windowed
+from more_itertools import windowed, flatten
 
 from renard.pipeline.ner import NEREntity
 from renard.pipeline.core import PipelineStep
@@ -15,12 +14,12 @@ from renard.pipeline.quote_detection import Quote
 
 def sent_index_for_token_index(token_index: int, sentences: List[List[str]]) -> int:
     """Compute the index of the sentence of the token at ``token_index``"""
-    sents_len = accumulate([len(s) for s in sentences], operator.add)
+    sents_len = it.accumulate([len(s) for s in sentences], operator.add)
     return next((i for i, l in enumerate(sents_len) if l > token_index))
 
 
-def sent_indices_for_blocks(
-    dynamic_blocks: List[List[str]], block_idx: int, sentences: List[List[str]]
+def sent_indices_for_block(
+    dynamic_blocks: List[List[str]], block_i: int, sentences: List[List[str]]
 ) -> Tuple[int, int]:
     """Return the indices of the first and the last sentence of a
     block
@@ -31,23 +30,21 @@ def sent_indices_for_blocks(
     :param sentences: all sentences
     :return: ``(first sentence index, last sentence index)``
     """
-    block_start_idx = sum(
-        [len(c) for i, c in enumerate(dynamic_blocks) if i < block_idx]
-    )
-    block_end_idx = block_start_idx + len(dynamic_blocks[block_idx])
-    sents_start_idx = None
-    sents_end_idx = None
+    block_start = sum([len(c) for i, c in enumerate(dynamic_blocks) if i < block_i])
+    block_end = block_start + len(dynamic_blocks[block_i])
+    sents_start = None
+    sents_end = None
     count = 0
     for sent_i, sent in enumerate(sentences):
-        start_idx, end_idx = (count, count + len(sent))
-        count = end_idx
-        if sents_start_idx is None and start_idx >= block_start_idx:
-            sents_start_idx = sent_i
-        if sents_end_idx is None and end_idx >= block_end_idx:
-            sents_end_idx = sent_i
+        start, end = (count, count + len(sent))
+        count = end
+        if sents_start is None and start >= block_start:
+            sents_start = sent_i
+        if sents_end is None and end >= block_end:
+            sents_end = sent_i
             break
-    assert not sents_start_idx is None and not sents_end_idx is None
-    return (sents_start_idx, sents_end_idx)
+    assert not sents_start is None and not sents_end is None
+    return (sents_start, sents_end)
 
 
 def mentions_for_blocks(
@@ -65,7 +62,7 @@ def mentions_for_blocks(
     blocks_mentions = [[] for _ in range(len(dynamic_blocks))]
 
     start_indices = list(
-        itertools.accumulate([0] + [len(block) for block in dynamic_blocks[:-1]])
+        it.accumulate([0] + [len(block) for block in dynamic_blocks[:-1]])
     )
     end_indices = start_indices[1:] + [start_indices[-1] + len(dynamic_blocks[-1])]
 
@@ -85,13 +82,10 @@ class CoOccurrencesGraphExtractor(PipelineStep):
         self,
         co_occurrences_dist: Optional[
             Union[int, Tuple[int, Literal["tokens", "sentences"]]]
-        ],
+        ] = None,
         dynamic: bool = False,
         dynamic_window: Optional[int] = None,
         dynamic_overlap: int = 0,
-        co_occurences_dist: Optional[
-            Union[int, Tuple[int, Literal["tokens", "sentences"]]]
[…]
