+++ _sources/introduction.rst.txt
@@ -22,16 +22,3 @@ networks and their automated extraction.
 
 .. _LabatutAndBost2019: https://arxiv.org/abs/1907.02704
 
-
-Target Audience and Intended Usage
-==================================
-
-Renard can be used by digital humanities researchers to extract
-character networks, for example to bring additional information when
-performing literary analysis.
-
-Renard can also be used by Natural Language Processing (NLP)
-researchers and practitioners to extract graph representations of
-texts, in order to use these as inputs to downstream NLP tasks. Graphs
-embedding algorithms such as Graph2Vec may be used on these graphs to
-derive a vectorial representation of texts.
