### 2024-05-21 10b2a80
progress toward dynamic_blocks+co_occurrences_blocks support
