+++ renard/pipeline/tokenization.py
@@ -1,49 +1,78 @@
-from typing import Dict, Any, List, Optional, Set, Union, Literal
-import itertools
-import torch
+from typing import Dict, Any, Set, Union, Literal, List, Tuple
+from more_itertools import windowed
 import nltk
-from transformers.tokenization_utils_base import BatchEncoding
-from more_itertools.recipes import flatten
+from nltk.data import load
+from nltk.tokenize.destructive import NLTKWordTokenizer
 from renard.pipeline.core import PipelineStep
-from renard.pipeline.progress import ProgressReporter
 from renard.nltk_utils import NLTK_ISO_STRING_TO_LANG
 
 
+def make_char2token(text: str, token2chars: List[Tuple[int, int]]) -> List[int]:
+    if len(token2chars) == 0:
+        return []
+
+    c2t = [None] * len(text)
+    for token_i, chars in enumerate(token2chars):
+        for char_i in range(*chars):
+            c2t[char_i] = token_i  # type: ignore
+
+    for char_i in range(0, token2chars[0][0]):
+        c2t[char_i] = 0  # type: ignore
+    for chars1, chars2 in windowed(token2chars, 2):
+        if chars1 is None or chars2 is None:
+            continue
+        end1 = chars1[1]
+        start2 = chars2[0]
+        for char_i in range(end1, start2):
+            c2t[char_i] = c2t[end1 - 1]
+    for char_i in range(token2chars[-1][1], len(c2t)):
+        c2t[char_i] = token2chars[-1][1]  # type: ignore
+
+    assert all([not i is None for i in c2t])
+    return c2t  # type: ignore
+
+
 class NLTKTokenizer(PipelineStep):
-    """Construct a nltk word tokenizer"""
+    """A NLTK-based tokenizer"""
 
     def __init__(self):
         nltk.download("punkt", quiet=True)
+        self.word_tokenizer = None
+        self.sent_tokenizer = None
         super().__init__()
 
-    def __call__(
-        self, text: str, dynamic_blocks: Optional[List[str]] = None, **kwargs
-    ) -> Dict[str, Any]:
-        """
-        :param text:
-        """
-        if not dynamic_blocks is None:
-            out_dicts = [self.__call__(block, None) for block in dynamic_blocks]
-            return {
-                "tokens": list(itertools.chain(*[d["tokens"] for d in out_dicts])),
-                "sentences": list(
-                    itertools.chain(*[d["sentences"] for d in out_dicts])
-                ),
-                "dynamic_blocks_tokens": [d["tokens"] for d in out_dicts],
-            }
+    def _pipeline_init_(self, lang: str, **kwargs):
+        assert lang in NLTK_ISO_STRING_TO_LANG
+        nltk_lang = NLTK_ISO_STRING_TO_LANG[lang]
+        self.word_tokenizer = NLTKWordTokenizer()
+        self.sent_tokenizer = load(f"tokenizers/punkt/{nltk_lang}.pickle")
+        super()._pipeline_init_(lang, **kwargs)
 
-        sentences = nltk.sent_tokenize(
-            text, language=NLTK_ISO_STRING_TO_LANG[self.lang]
-        )
+    def __call__(self, text: str, **kwargs) -> Dict[str, Any]:
+        assert not self.word_tokenizer is None
+        assert not self.sent_tokenizer is None
+
+        sent_indices = self.sent_tokenizer.span_tokenize(text)
 
         tokens = []
+        token2chars = []
         tokenized_sentences = []
-        for sent in sentences:
-            sent_tokens = nltk.word_tokenize(sent)
+        for sent_start, sent_end in sent_indices:
+            sent = text[sent_start:sent_end]
+            sent_tokens_indices = list(self.word_tokenizer.span_tokenize(sent))
[…]
