+++ renard/pipeline/graph_extraction.py
@@ -4,12 +4,21 @@ import operator
 
 import networkx as nx
 import numpy as np
-from more_itertools import windowed, flatten
+from more_itertools import windowed
 
 from renard.pipeline.ner import NEREntity
 from renard.pipeline.core import PipelineStep
 from renard.pipeline.character_unification import Character
 from renard.pipeline.quote_detection import Quote
+
+
+def char_blocks_to_token_blocks(
+    char_blocks: List[Tuple[int, int]], char2token: List[int]
+) -> List[Tuple[int, int]]:
+    tokens_blocks = []
+    for char_block_start, char_block_end in char_blocks:
+        tokens_blocks.append((char2token[char_block_start], char2token[char_block_end]))
+    return tokens_blocks
 
 
 def sent_index_for_token_index(token_index: int, sentences: List[List[str]]) -> int:
@@ -19,19 +28,15 @@ def sent_index_for_token_index(token_index: int, sentences: List[List[str]]) -> 
 
 
 def sent_indices_for_block(
-    dynamic_blocks: List[List[str]], block_i: int, sentences: List[List[str]]
+    dynamic_block: Tuple[int, int], sentences: List[List[str]]
 ) -> Tuple[int, int]:
     """Return the indices of the first and the last sentence of a
     block
 
-    :param dynamic_blocks: all blocks
-    :param block_idx: index of the block for which sentence
-        indices are returned
-    :param sentences: all sentences
+    :param dynamic_block: (START, END) in tokens
     :return: ``(first sentence index, last sentence index)``
     """
-    block_start = sum([len(c) for i, c in enumerate(dynamic_blocks) if i < block_i])
-    block_end = block_start + len(dynamic_blocks[block_i])
+    block_start, block_end = dynamic_block
     sents_start = None
     sents_end = None
     count = 0
@@ -41,6 +46,11 @@ def sent_indices_for_block(
         if sents_start is None and start >= block_start:
             sents_start = sent_i
         if sents_end is None and end >= block_end:
+            # this happens when the block is _smaller_ than the
+            # current sentence. In that case, we return the current
+            # sentence even though it overflows the block.
+            if sents_start is None:
+                sents_start = sent_i
             sents_end = sent_i
             break
     assert not sents_start is None and not sents_end is None
@@ -48,7 +58,7 @@ def sent_indices_for_block(
 
 
 def mentions_for_blocks(
-    dynamic_blocks: List[List[str]],
+    dynamic_blocks: List[Tuple[int, int]],
     mentions: List[Tuple[Any, NEREntity]],
 ) -> List[List[Tuple[Any, NEREntity]]]:
     """Return each block mentions
@@ -61,13 +71,8 @@ def mentions_for_blocks(
     """
     blocks_mentions = [[] for _ in range(len(dynamic_blocks))]
 
-    start_indices = list(
-        it.accumulate([0] + [len(block) for block in dynamic_blocks[:-1]])
-    )
-    end_indices = start_indices[1:] + [start_indices[-1] + len(dynamic_blocks[-1])]
-
     for mention in mentions:
-        for block_i, (start_i, end_i) in enumerate(zip(start_indices, end_indices)):
+        for block_i, (start_i, end_i) in enumerate(dynamic_blocks):
             if mention[1].start_idx >= start_i and mention[1].end_idx < end_i:
                 blocks_mentions[block_i].append(mention)
                 break
@@ -127,6 +132,7 @@ class CoOccurrencesGraphExtractor(PipelineStep):
         if isinstance(co_occurrences_dist, int):
             co_occurrences_dist = (co_occurrences_dist, "tokens")
         self.co_occurrences_dist = co_occurrences_dist
+        self.need_co_occurrences_blocks = co_occurrences_dist is None
 
         if dynamic:
             if not dynamic_window is None:
[…]
