### 2024-05-29 #6574 @tellarin
Overall things look quite good. Just a couple comments...

### Functionality

**Installation:** I see the notice in the repo that the package supports only Python 3.8 to 3.10. Is there a specific reason to not support newer versions? I think the reason for this limitation should be documented in the repo, so future users can consider contributing to address it too.

BTW, the PyPI page for the package should also link back to the GitHub repo as the package homepage.

**Functionality:** The previous discussion clarified some questions I had. Thanks.

But I believe it would be good to say something more about the dynamic networks. The paper basically only covers the static scenario, but supporting dynamic networks is one of the highlighted differentiators of the package. One example and sentence about how using it looks like would be enough.

Also, when discussing the definitions of requirements for a pipeline step, it would be useful to point to somewhere in the repo for an example. This could be a footnote in the paper.

### Software paper

**References:**

The paper should cite Vala et al., 2015 as it's the base for the alias resolution currently in use.

It would also be interesting to cite the Labatut and Bost 2019 survey on character network extraction as background for the design of the system.

Lastly, capitalisation should be fixed in reference Rochat 2015. "... Émile Zola’s Les Rougon-Macquart."

**Quality of writing:** Below are some minor typos and other issues I've noticed:

* Line 10: "the user can" -> "users can"
* Line 12: "allows to specialize pipelines" -> "allows specializing"
* Line 15: More of a style issue, so it matches other places in the text. "that is, " -> "i.e., "
* Line 18: "actions..." -> "actions, etc."
* Line 19: missing comma after "i.e."
* Line 25: "and none of" -> "but none of"
* Line 48: "the concept of pipeline" -> "the concept of a pipeline"
* Line 69: "the user select" -> "users select"
* Line 69: "its custom pipeline" -> "their custom pipelines"
