### 2024-05-30 fd8c455
Add block_indices util function
