### 2024-05-30 02d4879
Add homepage, repository and documentation links to pyproject.toml
