### 2024-05-30 a4d345f
JOSS article: fix typos, add missing citations, fix citation casing
