+++ paper.bib
@@ -70,12 +70,11 @@
 
 @InProceedings{rochat-2015-character_networks_zola,
   author       = {Rochat, Y.},
-  title	       = {Character network analysis of Émile Zola’s Les
-                  Rougon-Macquart},
+  title	       = {Character network analysis of {Émile Zola’s Les
+                  Rougon-Macquart}},
   booktitle    = {Digital Humanities 2015},
   year	       = {2015},
-  url	       =
-                  {https://infoscience.epfl.ch/record/210573?ln=en},
+  url	       = {https://infoscience.epfl.ch/record/210573?ln=en},
 }
 
 @InProceedings{hagberg-2008-networkx,
@@ -85,4 +84,13 @@
   booktitle    = {7th Python in Science Conference (SciPy2008)},
   year	       = {2008},
   url	       = {https://www.osti.gov/biblio/960616},
+}
+
+@InProceedings{vala-2015-character_detection,
+  author    = {Vala, H. and Jurgens, D. and Piper, A. and Ruths, D.},
+  booktitle = {Conference on Empirical Methods in Natural Language Processing},
+  title     = {Mr. Bennet, his coachman, and the Archbishop walk into a bar but only one of them gets recognized: On The Difficulty of Detecting Characters in Literary Texts},
+  year      = {2015},
+  pages     = {769-774},
+  doi       = {10.18653/v1/D15-1088},
 }
