### 2024-05-30 24b8f65
Update documentation to add more details on requirements errors (+ small fixs)
