### 2024-05-30 703c376
JOSS article: add some missing details, more links to documentation
