+++ paper.md
@@ -32,7 +32,7 @@ Renard (*Relationships Extraction from NARrative Documents*) is a Python library
 
 Character networks (i.e., graphs where vertices represent characters and edges represent their relationships) extracted from narrative texts are useful in a number of applications, from visualization to literary analysis [@labatut-2019]. There are different ways of modeling relationships (co-occurrences, conversations, actions, etc.), and networks can be static or dynamic (i.e., series of networks representing the evolution of relationships through time). This variety means one can extract different kinds of networks depending on the targeted applications. While some authors extract these networks by relying on manually annotated data [@rochat-2014-phd_thesis; @rochat-2015-character_networks_zola; @rochat-2017; @park-2013-character_networks; @park-2013-character_networksb], it is a time-costly endeavor, and the fully automatic extraction of these networks is therefore of interest. Unfortunately, there are only a few existing software and tools that can extract character networks [@marazzato-2014-chaplin; @metrailler-2023-charnetto], but none of these can output dynamic networks. Furthermore, automatically extracting a character network requires solving several successive natural language processing tasks, such as named entity recognition (NER) or coreference resolution, and algorithms carrying these tasks are bound to make errors. To our knowledge, the cascading impact of these errors on the quality of the extracted networks has yet to be studied extensively. This is an important issue since knowing which tasks have more influence on the extracted networks would allow prioritizing research efforts.
 
-Renard is a fully configurable pipeline that can extract static and dynamic networks from narrative texts. We base Renard on the generic character network extraction framework highlighted by the survey of [@labatut-2019]. We design it so that it is as modular as possible, which allows the user to select the implementation of each extraction step as needed. This has several advantages:
+Renard is a fully configurable pipeline that can extract static and dynamic networks from narrative texts. We base Renard on the generic character network extraction framework highlighted by the survey of @labatut-2019. We design it so that it is as modular as possible, which allows the user to select the implementation of each extraction step as needed. This has several advantages:
 
 1. Depending on the input text, the user can choose the most relevant series of steps and configure each of them as needed. Therefore, the pipeline can be specialized for different types of texts, allowing for better performance.
 2. The pipeline can easily incorporate new advances in NLP, by simply implementing a new step when necessary.
@@ -69,7 +69,9 @@ out = pipeline(text)
 
 ![Co-occurrence character network of Jane Austen's "Pride and Prejudice", extracted automatically using Renard. Vertex size and color denote degree, while edge thickness and color denote the number of co-occurrences between two characters.](./pp.pdf){#fig:pp_network height=30% }
 
-As an example, Figure \ref{fig:pp_network} shows the co-occurrence character network Jane Austen's 1813 novel "Pride and Prejudice", extracted using the Renard pipeline above. Renard uses the NetworkX Python library [@hagberg-2008-networkx] to manipulate graphs, ensuring compatibility with a wide array of tools and formats.
+As an example, Figure \ref{fig:pp_network} shows the co-occurrence character network Jane Austen's 1813 novel "Pride and Prejudice", extracted using the Renard pipeline above. While this network is static, users can also extract dynamic networks by passing the `dynamic=True` argument to the last step of the pipeline: in that case, Renard outputs a list of graphs corresponding to a dynamic network instead of a single network[^1]. Renard uses the NetworkX Python library [@hagberg-2008-networkx] to manipulate graphs, ensuring compatibility with a wide array of tools and formats.
+
+[^1]: See [the documentation on dynamic networks](https://compnet.github.io/Renard/pipeline.html#dynamic-graphs) for more details.
 
 To allow for custom needs, we design Renard to be very flexible. If a step is not available in Renard, we encourage users to either:
 
[…]
