### 2024-05-30 16a3aa5
Add more details to the Dynamic Graphs section of the documentation
