+++ _sources/pipeline.rst.txt
@@ -22,7 +22,7 @@ document. Here is a simple example:
        [
            NLTKTokenizer(),
            NLTKNamedEntityRecognizer(),
-           GraphRulesCharacterUnifier(min_appearance=10),
+           GraphRulesCharacterUnifier(min_appearances=10),
            CoOccurrencesGraphExtractor(co_occurrences_dist=25)
        ]
    )
@@ -55,13 +55,20 @@ to compute them yourself:
    pipeline = Pipeline(
        [
            NLTKNamedEntityRecognizer(),
-           NaiveCharactersExtractor(min_appearance=10),
+           GraphRulesCharacterUnifier(min_appearances=10),
            CoOccurrencesGraphExtractor(co_occurrences_dist=25)
        ]
    )
 
    # tokens are passed at call time
    out = pipeline(text, tokens=my_tokenization_function(text))
+
+
+In that case, the ``tokens`` requirements is fulfilled at run time. If
+you don't pass the parameter, Renard will throw the following
+exception:
+
+>>> ValueError: ["step 1 (NLTKNamedEntityRecognizer) has unsatisfied needs (needs : {'tokens'}, available : {'text'})"]
 
 
 For simplicity, one can use one of the preconfigured pipelines:
@@ -232,6 +239,9 @@ When executing the above block of code, the output attribute
 
 >>> out.character_network
 [<networkx.classes.graph.Graph object at 0x7fd9e9115900>]
+
+See :class:`.CoOccurrencesGraphExtractor` for more details on the
+usage of the ``dynamic`` and ``dynamic_window`` arguments.
 
 Plot and export functions work as one would expect
 intuitively. :meth:`.PipelineState.plot_graph` allow to visualize the
