### 2024-05-30 cba9070
JOSS article: add more details on dynamic networks extraction
