### 2024-05-30 #6574 @Aethor
Hi,

thank you for reviewing Renard!

>Installation: I see the notice in the repo that the package supports only Python 3.8 to 3.10. Is there a specific reason to not support newer versions? I think the reason for this limitation should be documented in the repo, so future users can consider contributing to address it too.

There is no specific reason for not supporting Python 3.11+, aside from maintenance time. We do update the supported Python versions from time to time, but we always have to fix some issues with Renard's dependencies when we do so. We hope to support Python 3.11 soon!

>BTW, the PyPI page for the package should also link back to the GitHub repo as the package homepage.

Good point, we fixed this by adding the relevant links in the `pyproject.toml` file (see 02d4879a2eb6580265462165affbb51c46f63320). As a result, we published Renard 0.4.1 to update the project on PyPi (and ship a few bugfixs).

>But I believe it would be good to say something more about the dynamic networks. The paper basically only covers the static scenario, but supporting dynamic networks is one of the highlighted differentiators of the package. One example and sentence about how using it looks like would be enough.

We added a new sentence explaining how Renard could be used to extract a dynamic network, and a footnote with a link to the section of the documentation specifically related to extracting dynamic networks. We also modified the example code in the article to show where to pass parameters to extract a dynamic network.

>Also, when discussing the definitions of requirements for a pipeline step, it would be useful to point to somewhere in the repo for an example. This could be a footnote in the paper.

We added a footnote in the paper, pointing to the documentation. In the documentation itself, we added an example of what would happen in case of a missing step requirement (see 24b8f6525fe039118bde424425f492fe8748a0e1).

>The paper should cite Vala et al., 2015 as it's the base for the alias resolution currently in use.
>It would also be interesting to cite the Labatut and Bost 2019 survey on character network extraction as background for the design of the system.

We modified the article to include these citations. We now state that *"we base Renard on the generic character network extraction framework highlighted by the survey of Labatut and Bost (2019)"*.

>Lastly, capitalisation should be fixed in reference Rochat 2015. "... Émile Zola’s Les Rougon-Macquart."

>Quality of writing: Below are some minor typos and other issues I've noticed:

Thanks, we fixed these as well!
