### 2024-06-04 #6574 @logological
@Aethor Where possible, could you please update your paper to supply the missing DOIs reported by editorialbot above?  Possibly the only reference that has a DOI that is not given is Marazzato & Sparavigna: 10.48550/arXiv.1402.4259.  However, you might consider replacing this reference, an unreviewed preprint, with what what seems to be the later journal paper based on it:

> Sparavigna, Amelia Carolina and Marazzato, Roberto (March 2015). Analysis of a Play by Means of CHAPLIN, the Characters and Places Interaction Network Software. _International Journal of Sciences_, 2015, 4(3):60-68. DOI: [10.18483/ijSci.662](http://dx.doi.org/10.18483/ijSci.662)
