### 2024-06-05 d97e2fe
JOSS article: update ref marazzato-2014-chaplin to sparavigna-2015-chaplin
