+++ paper.bib
@@ -1,19 +1,21 @@
 ;;; -*- eval: (bibtex-set-dialect 'biblatex); -*-
-@Article{marazzato-2014-chaplin,
-  title	       = {Extracting Networks of Characters and Places from
-                  Written Works with CHAPLIN},
-  author       = {Marazzato, R. and Sparavigna, A. C.},
-  year	       = {2014},
-  volume       = {cs.CY},
-  pages	       = {1402.4259},
-  journal      = {arXiv},
+@Article{sparavigna-2015-chaplin,
+  title	       = {Analysis of a Play by Means of CHAPLIN, the
+                  Characters and Places Interaction Network Software},
+  author       = {Sparavigna, A. C. and Marazzato, R.},
+  journal      = {International Journal of Sciences},
+  year	       = {2015},
+  doi	       = {10.18483/ijSci.662},
+  volume       = {4},
+  number       = {3},
+  pages	       = {60-68},
 }
 
 @Misc{metrailler-2023-charnetto,
-  title = {{Charnetto}},
-  url = {https://gitlab.com/maned_wolf/charnetto},
-  author = {Métrailler, C.},
-  year = {2023},
+  title	       = {{Charnetto}},
+  url	       = {https://gitlab.com/maned_wolf/charnetto},
+  author       = {Métrailler, C.},
+  year	       = {2023},
 }
 
 @Article{labatut-2019,
