### 2024-06-05 #6574 @Aethor
Good catch, Sparavigna and Marazzato (2015) also presents their CHAPLIN software so I updated the reference to cite this journal article instead of their earlier arXiv preprint. I also did another pass to confirm that the other references missing DOIs really did not have one, and I was unable to find any additional DOI.
