### 2024-06-12 cb340c4
BertCoreferenceResolver: code simplification, support for Tibert's hierarchical merging
