### 2024-06-13 2653354
Fix BertCoreferenceResolver wrong output when not using hierarchical_merging
