### 2024-06-17 #6574 @LCB0B
@logological @Aethor The changes look good to me as well. I'm happy with the current state of the manuscript.
