### 2024-06-19 a144119
Add ignore_lone_titles option for GraphRulesCharacterUnifier
