+++ _sources/pipeline.rst.txt
@@ -22,7 +22,7 @@ document. Here is a simple example:
        [
            NLTKTokenizer(),
            NLTKNamedEntityRecognizer(),
-           GraphRulesCharacterUnifier(min_appearance=10),
+           GraphRulesCharacterUnifier(min_appearances=10),
            CoOccurrencesGraphExtractor(co_occurrences_dist=25)
        ]
    )
@@ -55,13 +55,20 @@ to compute them yourself:
    pipeline = Pipeline(
        [
            NLTKNamedEntityRecognizer(),
-           NaiveCharactersExtractor(min_appearance=10),
+           GraphRulesCharacterUnifier(min_appearances=10),
            CoOccurrencesGraphExtractor(co_occurrences_dist=25)
        ]
    )
 
    # tokens are passed at call time
    out = pipeline(text, tokens=my_tokenization_function(text))
+
+
+In that case, the ``tokens`` requirements is fulfilled at run time. If
+you don't pass the parameter, Renard will throw the following
+exception:
+
+>>> ValueError: ["step 1 (NLTKNamedEntityRecognizer) has unsatisfied needs (needs : {'tokens'}, available : {'text'})"]
 
 
 For simplicity, one can use one of the preconfigured pipelines:
@@ -233,6 +240,9 @@ When executing the above block of code, the output attribute
 >>> out.character_network
 [<networkx.classes.graph.Graph object at 0x7fd9e9115900>]
 
+See :class:`.CoOccurrencesGraphExtractor` for more details on the
+usage of the ``dynamic`` and ``dynamic_window`` arguments.
+
 Plot and export functions work as one would expect
 intuitively. :meth:`.PipelineState.plot_graph` allow to visualize the
 dynamic graph using a slider, and
@@ -240,49 +250,6 @@ dynamic graph using a slider, and
 graph to a directory. Meanwhile,
 :meth:`.PipelineState.export_graph_to_gexf` correctly exports the
 dynamic graph to the Gephi format.
-
-
-Custom Segmentation
--------------------
-
-The ``dynamic_window`` parameter of
-:class:`.CoOccurencesGraphExtractor` determines the segmentation of
-the dynamic networks, in number of interactions. In the example above,
-a new graph will be created for each 20 interactions.
-
-While one can rely on the arguments of the graph extractor of the
-pipeline to determine the dynamic window, Renard allows to specify a
-custom segmentation of a text with the ``dynamic_blocks``
-argument. When running a pipeline, you can cut your text however you
-want and pass this argument instead of the usual text:
-
-
-.. code-block:: python
-
-   from renard.pipeline import Pipeline
-   from renard.pipeline.tokenization import NLTKTokenizer
-   from renard.pipeline.ner import NLTKNamedEntityRecognizer
-   from renard.pipeline.character_unification import GraphRulesCharacterUnifier
-   from renard.pipeline.graph_extraction import CoOccurrencesGraphExtractor
-   from renard.utils import blocks_indices
-
-   with open("./my_doc.txt") as f:
-       text = f.read()
-
-   # let's suppose the 'cut_into_chapters' function cut the text into chapters
-   chapters = cut_into_chapters(text)
-
-   pipeline = Pipeline(
-       [
-           NLTKTokenizer(),
-           NLTKNamedEntityRecognizer(),
-           GraphRulesCharacterUnifier(min_appearances=10),
-           CoOccurrencesGraphExtractor(co_occurrences_dist=25, dynamic=True)
-       ]
-   )
-
-   out = pipeline(dynamic_blocks=block_indices(chapters))
-
 
 
 Multilingual Support
