### 2024-06-19 bc9b340
bump version to 0.4.2
