### 2024-06-19 #6574 @Aethor
- I double checked that the ORCIDs and affiliations in the paper are correct
- The latest Renard release is now `0.4.2` - see https://github.com/CompNet/Renard/releases/tag/v0.4.2
- We archived the release on Zenodo with the following DOI: 10.5281/zenodo.12167899 - see https://zenodo.org/records/12167900
- I double checked the title and authors list on Zenodo and I can confirm that it is correct
- The license of the archive is the same as Renard : GPLv3
