### 2024-06-19 #6574 @logological
> * The license of the archive is the same as Renard : GPLv3

@Aethor I'm afraid this is not correct.  Your GitHub repository includes the GPLv3 licence in the `LICENSE` file and the `pyproject.toml` file indicates the licence is `GPL-3.0-only`. However, in Zenodo you have marked the licence as "GNU General Public License v3.0 or later" (i.e., `GPL-3.0-or-later`).  Could you please reconcile this?  You should either change your `pyproject.toml` so that it indicates `GPL-3.0-or-later` (provided all contributors holding copyright agree to this) or else change the licence tag on Zenodo.
