### 2024-06-19 #6574 @Aethor
@logological oh, you're right! I fixed the license in Zenodo, so it is now  GNU General Public License v3.0 only.
