### 2024-06-19 2c5780a
Copy-editing: 'allow' must be transitive here
