### 2024-06-19 #6574 @logological
@Aethor Thanks for fixing the licence tag!  In other news, I have identified a few grammatical errors in the paper and submitted a pull request.  Could you please review it and (if you are happy with it) accept it?
