### 2024-06-19 #6574 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
