### 2024-06-20 #6574 @crvernon
Nice work on this @Aethor!  I just need you to make the following change to the paper before I can accept this for publication (no need to make a new release after changing the paper documents):

- [ ] LINE 56 "...network Jane Austen’s..." should be "...network of Jane Austen’s..."

Thanks!
