### 2024-06-20 f242c6e
Fix missing "of" in a sentence of the JOSS paper
